chore: move types to cosmos.d.ts
